修改单一任务下载完成后没有删除历史任务的bug

+++ downloadlib/src/main/java/cn/can/downloadlib/DownloadManager.java
@@ -687,8 +687,10 @@ public class DownloadManager implements AppInstallListener {
     }
 
     public void deleteSigleTask(String taskId) {
-        mSingleTaskMap.remove(taskId);
-        mDownloadDao.deleteByKey(taskId);
+        if (mSingleTaskMap != null && taskId != null) {
+            mSingleTaskMap.remove(taskId);
+            mDownloadDao.deleteByKey(taskId);
+        }
     }
 
     public void setLimitSpace(int size) {
